+++ CHANGELOG.md
@@ -1,4 +1,25 @@
-# March 20th, 2016
+# DATE
+
+## Interpreter
+
+0. Support of indexers.
+0. Bug fixed with the step into.
+0. Refactoring.
+0. Optimization.
+
+## Parser
+
+0. Support of program declaration.
+0. Support of class declaration.
+0. Support of method declaration.
+0. Support of variable declaration (but not with default value).
+
+### Know issues
+
+0. As the parser is not complete, it does not works correctly.
+0. The interpreter is still not able to access the UI.
+
+# March 20th, 2016 (branch dev)
 
 0. Beginning of a programming language parser. The goal is to be able to type a code, generate an AlgorithmProgram and interpret it.
 0. Support of variables size trace removed, because it has been judged useless. The dependence BareMetal of this feature has been removed.
@@ -10,7 +31,7 @@
 0. As the parser is not complete, it does not works correctly.
 0. The interpreter is still not able to access the UI.
 
-# March 13th, 2016
+# March 13th, 2016 (branch master)
 
 0. When the algorithm interpreter is in pause in debug mode, we can now perform a Step Into.
 0. Step Over.
@@ -21,29 +42,29 @@
 0. Bug that show less log than excpected with iteration and condition fixed.
 0. Know performance issue => when using operators, the interpreter is quite slower at first run than previous check-in.
 
-# March 12th, 2016
+# March 12th, 2016 (branch master)
 
 0. We can now pause the algorithm interpreter.
 0. We can add a breakpoint to an algorithm with AlgorithmBreakpointStatement.
 0. We can resume a paused (by breakpoint or not) algorithm interpreter.
 0. Bug fixed.
 
-# March 9th, 2016
+# March 9th, 2016 (branch master)
 
 0. Stop any algorithm at runtime even with an infinite loop.
 0. Keep a call stack and the values of each accessibles variables.
 0. Refactoring.
 0. Performance improvement.
 
-# March 7th, 2016
+# March 7th, 2016 (branch master)
 
 0. StackOverflowException with recursivity avoided. 
 
-# March 6th, 2016
+# March 6th, 2016 (branch master)
 
 0. Performance improvement with Binary Operator.
 0. Bug fixed.
 
-# April 19th, 2015
+# April 19th, 2015 (branch master)
 
 0. First prototype.